[HOT-FIX] remove more compiler warnings
Former-commit-id: 9cd7195fdf1c6a765399c55beb23671ed580895d
Former-commit-id: a37bed7abf45ea7622f068439dd2d1d1a16d80ba

+++ src/simulation/onboardDataHandling/spaceToGroundTransmitter/spaceToGroundTransmitter.cpp
@@ -122,7 +122,7 @@ void SpaceToGroundTransmitter::evaluateDataModel(DataNodeUsageMsgPayload *dataUs
     // Get the buffer with the most data
     double maxVal = -1.0;
     int maxIndex = -1;
-    for (int i = 0; i < this->storageUnitMsgsBuffer.back().storedData.size(); i++) {
+    for (uint64_t i = 0; i < this->storageUnitMsgsBuffer.back().storedData.size(); i++) {
         if (this->storageUnitMsgsBuffer.back().storedData[i] > maxVal) {
             maxVal = this->storageUnitMsgsBuffer.back().storedData[i];
             maxIndex = i;
